+++ Assets/Scripts/IDamageable.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void takeDmg(int damage);
+    void takeDmg(float damage);
 }
